requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
pypdf
chromadb
python-dotenv
streamlit
tiktoken
